refactor(api): move acceptance test from /lib to /src/team
